I guess recording is better... but it's slow now?

+++ gui/qt/capture/animated-png.h
@@ -28,7 +28,7 @@ typedef struct {
 } apng_t;
 
 bool apng_start(const char *tmp_name, int frameskip);
-void apng_add_frame(const void *frame, int num, int den);
+void apng_add_frame(const void *frame, uint16_t num, uint16_t den);
 bool apng_stop(void);
 bool apng_save(const char *filename, bool optimize);
